feat: alerts and spinners Home Page
